Moved browser cleanup to ConcordionBase
