--- src/test/java/example/ConcordionIndex.java
@@ -1,7 +1,0 @@
-package example;
-
-/**
- * Customise index specification.
- */
-public abstract class ConcordionIndex extends ConcordionDomainBase {
-}
