+++ src/test/java/example/Example.java
@@ -3,6 +3,6 @@ package example;
 /**
  * Run me as a Junit Test to test all specifications.
  */
-public class Example extends ConcordionIndex {
+public class Example extends CubanoDemoIndex {
 
 }
