+++ src/test/java/example/commands/SetCommandFixture.java
@@ -1,8 +1,8 @@
 package example.commands;
 
-import example.ConcordionFixture;
+import example.CubanoDemoFixture;
 
-public class SetCommandFixture extends ConcordionFixture {
+public class SetCommandFixture extends CubanoDemoFixture {
 
     public String hello(String name) throws Exception {
