+++ src/test/java/example/cubanoconcordion/ToolTipFixture.java
@@ -1,13 +1,12 @@
 package example.cubanoconcordion;
 
-import example.ConcordionFixture;
+import example.CubanoDemoFixture;
 
-public class ToolTipFixture extends ConcordionFixture {
+public class ToolTipFixture extends CubanoDemoFixture {
 
     public boolean addToolTipLogging() {
 
         addConcordionTooltip("Look at me, just hovering over here!");
         return true;
     }
-
 }
