+++ src/test/java/example/cubanocore/UserPoolFixture.java
@@ -3,11 +3,13 @@ package example.cubanocore;
 import org.concordion.cubano.template.driver.domain.Role;
 import org.concordion.cubano.template.driver.domain.User;
 
-import example.ConcordionFixture;
+import example.CubanoDemoBrowserFixture;
+import org.concordion.cubano.template.driver.domain.UserPool;
 
-public class UserPoolFixture extends ConcordionFixture {
+public class UserPoolFixture extends CubanoDemoBrowserFixture {
 
     private User user;
+    private UserPool userPoolManager;
 
     public String grabUser() {
 
@@ -37,4 +39,19 @@ public class UserPoolFixture extends ConcordionFixture {
         return !userPool().userIsInUse(user);
     }
 
+    protected User getUser() {
+        if (user == null) {
+            user = userPool().requestUser(Role.ROLE1);
+        }
+
+        return user;
+    }
+
+    public UserPool userPool() {
+        if (userPoolManager == null) {
+            userPoolManager = UserPool.createManager(getCleanupService());
+        }
+
+        return userPoolManager;
+    }
 }
