+++ src/test/java/example/cubanohttpeasy/HttpEasyPostFixture.java
@@ -7,6 +7,7 @@ import static com.github.tomakehurst.wiremock.client.WireMock.urlPathMatching;
 
 import java.io.IOException;
 
+import example.CubanoDemoBrowserFixture;
 import org.concordion.cubano.driver.http.HttpEasy;
 import org.concordion.cubano.driver.http.JsonReader;
 import org.concordion.cubano.template.driver.logger.TestLoggerLogWriter;
@@ -20,9 +21,7 @@ import org.junit.Rule;
 import com.github.tomakehurst.wiremock.junit.WireMockRule;
 import com.google.common.net.MediaType;
 
-import example.ConcordionFixture;
-
-public class HttpEasyPostFixture extends ConcordionFixture {
+public class HttpEasyPostFixture extends CubanoDemoBrowserFixture {
 
     private static final String HTTP_EASY_POST = "/HttpEasyPost";
