+++ src/test/java/example/google/SearchForConcordionFixture.java
@@ -1,11 +1,13 @@
 package example.google;
 
-import example.ConcordionFixture;
+import example.CubanoDemoBrowserFixture;
+import org.concordion.cubano.template.driver.workflow.Workflow;
 
-public class SearchForConcordionFixture extends ConcordionFixture {
+public class SearchForConcordionFixture extends CubanoDemoBrowserFixture {
+    private Workflow workflow = new Workflow(this);
 
     public String google(String term, String site) {
-        return workflow()
+        return workflow
                 .openSearch()
                 .searchFor(term)
                 .getSearchResult(site);
